requirements.txt
certifi==2018.8.24
chardet==3.0.4
dj-database-url==0.5.0
Django==1.11.15
gunicorn==19.9.0
idna==2.7
psycopg2==2.7.5
pytz==2018.5
requests==2.19.1
urllib3==1.23
whitenoise==3.3.1
djangorestframework==3.9.0
